Fix subsequent UpsertServer() calls for memng
The `UpsertServer()` method for the `memng` engine was only storing the
first server for a backend. Subsequent upserts for additional servers were
ignored because `vals` wasn't written back to `Mem.Servers`. I have updated
the tests API tests, which use memng, to exercise this.

Technically updating servers with `vals[i] = srv` was doing the right thing
but I've changed that too for better readability.