Remove backend from backends array, when it goes out of ETCD
During our high-frequency change scenario (we use vulcand programmatically),
we discovered a weird edge condition on backends still living in the backends
array and thus causing old servers to be hit.

The crux of the problem was this:
1. PUT /backends/b1/servers/s1
2. PUT /frontends/f1/frontend {backendId: b1}
3. DELETE /frontends/f1
4. DELETE /backends/b1
5. PUT /backends/b1/servers/s2
6. PUT /frontnends/f1/frontned {backendId: b1}

Now if we hit frontend f1, we mysteriously saw hits to server s1.

Upon deeper investigation I found that in the mux.go:DeleteBackend function,
the backend was never really REMOVED from the m.backends map.

This meant that when upsertBackend is called, m.backends[bk] retrieves
the old backend, which is "updated" with the new backend, instead
of a completely new backend being created.

This left over some servers in the backend (either becuase etcd failed
to publish events or something went wrong.) Either way, if a frontend DID
hold a reference, from an outside observer, merging of a new etcd backend
with an old one makes little sense. AKA

1. PUT /backends/b1/servers/s1
2. PUT /frontends/f1/frontend {backendId: b1}
3. DELETE /backends/b1
4. PUT /backends/b1/servers/s2
5. PUT /frontneds/f2/frontend {backendId: b1}

While I totally admit this scenario is illegal, I DO believe
that frontend f2 should get the backend that exists in etcd at the time
the frontend was created. It cannot be expected to have memory of what
other users may or may not have done in the past. Basically, if I had
a worker process, it should be able to delete everything
and reset it from scratch to get the state it needs.

By explicitly removing that backend from the m.backends map, it means
if a frontend f2 is added without FIRST having added a backend, a proper
error will be raised, and if a backend did exist, THAT is the latest
backend it would hold a reference to.

A note about the DeferClose:
1. It is a simple form of reference-counting. When the backend is deleted,
we tell it that it no longer is referenced.
2. When the last frontend that references it is unlinked, it will simply
close its idle connections.