Revendor oxy again, and utilize the PassHostHeader option
While reviewing the oxy code, I realized the PassHostHeader optSetter
was not using its bool parameter. I made a change to utilize that and
have a pull request pending upstream into oxy.

This change then takes that change and revendors oxy into vulcand.

Finally I enable passing PassHostHeader option to the Forwarder
by pulling it from HttpSettings structure in the frontend.

This should allow a frontend to pass the Host header back to
proxied endpoints depending on a configurable setting.