Added ability to sync cluster with Etcd periodically.
The purpose of this change is to keep vulcand resilient against
long-running etcd clusters. This change allows vulcan to periodically
synchronize the latest members of the etcd cluster from the
member it is current connected to, this giving it fallback
members to use in case the primary member were to fail for
whatever reason.

This allows us to maintain the etcd cluster by adding new replacement machines,
and taking down older machines, but so long as the cluster itself
is running overall, vulcand instances talking to it, will continue to work.

Change seconds from integer type to time.Duration, so we can do time arithmetic on it.