Resolved a race condition.
This is what was happening:
1. Sending <- nil into a channel is the same as close()'ing the channel.
2. There were TWO producers onto the channel restartC who both wanted to tell supervise() to stop what it is doing.
3. The supervisor.Stop() function closed the channel outright, while the goroutine which watched for engine.Subscribe() merely sent a nil into it.
4. Leaving the channel open, however, prevents some other process downstream from proceeding.

What this meant was, under the right conditions, if Stop() was called while at the same time the engine was shutting down, the Stop()'s close(s.restartC) would end up closing the channel,
causing supervise() to break and beginning shutdown, at the same time, engine.Subscribe() would return with a nil error, and the go-routine would attempt to write nil onto a closed channel.

This happens very specifically in "TestGracefulShutdown" test. The test is correct, the code was the root cause.

This change does the following:
1. Under a mutex, both sites first check if the channel is non-nil and if so to do their own thing (the go routine, writes a nil to it, while Stop() just closes it.)
2. This ensures that no matter who gets there first, both of them perform just fine.
3. By setting the channel itself to "nil" it is a strong indication to all would-be writers that it does NOT exist.

I added delays in various places and ran the unit tests repeatedly under many different conditions. Everything seems to work reliably every single time.