Use channel broadcastCloseC to signal a Stop()
This change removes the original lock/mutex-based solution
to closing of channels. Instead, by adding a second channel which
can be written to by anyone, is never closed, and which, when written to,
ends up closing the restartC channel from a single place,
guarantees serialized proper behavior in all situations.