Add ability to disable Default Listener.
The context behind this change is security scenarios, or even
initializations scenarios where you don't want vulcan to
"leak" itself by accidentally exposing a listener that isn't ready
to handle connections.

Specifically under some situations, we may use vulcand as a middle
tier to white list what ports/hosts a connection may be possible to.

Under these situations, having a default listener exposes at least
one host/port we have to manage or hide.

This option defaults to "true" so it is 100% backwards compatible
and unless someone explicitly sets it to "false" you shouldn't
notice any difference. I chose to call it "DefaultListener"
as opposed to "NoDefaultListener" to avoid a double-negation.

It's a lot easier to read:
if options.DefaultListener {...

instead of

if !options.NoDefaultListener {...