Rename connTracker to incomingConnTracker
So I kinda messed up. I wanted to track connections going
OUT from vulcand not those coming in (necessarily), but this
provides a good opportunity to do both.

In the interest of not having ambiguity, I renamed the current connTracker
to "incomingConnTracker" or variants thereof as appropriate.

This opens the door to provide a second interface which can be passed back into oxy, that
allows tracking all outgoing connections being held open by vulcan.

As mentioned previously, the purpose of that is to allow us to
drain connections from servers before we kill them or replace them
without having to break a connected customer.