Fixed the GetSnapshot part of etcdv3ng
The UTs in suite don't cover the GetSnapshot function.

I'll write a comprehensive test for that later. Immediately though,
when vulcand starts with a pre-existing set of keys in etcd (over
clientv3), it doesn't pick up the snapshot correctly.

Also the function "filterKeysBySuffix" really filtered keys by
a common prefix. So renamed that.