Fixed the GetSnapshot part of etcdv3ng (#317)
* Allow specifying of flush interval from Frontend Settings when streaming

* I think it should be in nanoseconds just so that one-solution-wins-all.

* Made the interval int64 to match time.Duration's native type

* Go fmt

* Fixed the GetSnapshot part of etcdv3ng

The UTs in suite don't cover the GetSnapshot function.

I'll write a comprehensive test for that later. Immediately though,
when vulcand starts with a pre-existing set of keys in etcd (over
clientv3), it doesn't pick up the snapshot correctly.

Also the function "filterKeysBySuffix" really filtered keys by
a common prefix. So renamed that.