Provided a way to pass in more listeners to a frontend
Beginning with the Connection Tracker, which allowed
vulcan consumers to plug in customer tracking
functions when incoming connection states change,
there has been an increasing amount of rich information
that makes a lot of sense - either for debugging, or for
audition, to be emitted programmatically - in a manner
that is not tied to "log levels" where it can get
turned off accidentally.

This change simplifies the listeners into a consolidated
struct, so that they can be moved around and
wired through much more easily in the future.

This change adds two listeners - one for the
Rebalancer and one for RoundRobin.

The intention is to allow a consumer of vulcan to
have programmatic visibility into the rebalancing and
round robin decisions made.

This is also an easier mechanism to debug rewrite implications
rather than tracing logs.