Fix SNI Test to work with multiple hosts correctly (#344)
* Fix SNI to work with multiple hosts correctly

When testing AutoCert, I realized that this test was not comprehensive enough.

If we add an additional host call, say "foobar", we get a 404 back because the
route doesn't exist for it. However, the default cert is returned regardless.

This means that if we checked for the Certificate Serial Number on the returned
cert, it is constant across any host name. In short, exposing multiple SSL
Certs by Host, will not result in different certs being returned
based on the request made.

I have set up the initial conditions for the test which fails right now.

I'll figure out how to make it pass.

* Fixed SNI by setting client-side ServerName in each TLS request.

* Fixed SNI by setting client-side ServerName in each TLS request.

* Updated SNI test to be better formed

* Removed unnecessary testing in SNI, and lowercased the getPeerCertSerialNo