AutoCert (Let's Encrypt) Support. (#339)
* Fix SNI to work with multiple hosts correctly

When testing AutoCert, I realized that this test was not comprehensive enough.

If we add an additional host call, say "foobar", we get a 404 back because the
route doesn't exist for it. However, the default cert is returned regardless.

This means that if we checked for the Certificate Serial Number on the returned
cert, it is constant across any host name. In short, exposing multiple SSL
Certs by Host, will not result in different certs being returned
based on the request made.

I have set up the initial conditions for the test which fails right now.

I'll figure out how to make it pass.

* Fixed SNI by setting client-side ServerName in each TLS request.

* Fixed SNI by setting client-side ServerName in each TLS request.

* Updated SNI test to be better formed

* AutoCert Support (Let's Encrypt)

This change enables Let's Encrypt support through
use of the autocert package provided by golang.

Key things of Note:
1. A Host now has "AutoCert" as a field, in addition to KeyPair.
   When KeyPair is set, that key takes precedence to AutoCert.

   When AutoCert is set, it attempts an automatic Certificate
   generation for that hostname.

2. AutoCert field itself has a few members, namely:
   a) A DirectoryURL that allows you to set what ACME service to
      get certificates from (the ACME CA). Defaults to Let's Encrypt.

   b) A Key that is a PEM-encoded Private Key which authenticates
      you as a certificate-requestor to your ACME CA.
      A new 256-bit EC Key is generated if one is not provided.

   c) An Email that the CA can email you at, for failures or errors.

3. By default the Engines provide an AutoCertCache (where automatically
   generated certificates are stored.) This allows certificates to be
   cached along with Vulcan's own routing state.

   For instance, Etcd Engines will cache certificates in Etcd.

4. Caching is made a bit generic by providing an interface called
   a CacheProvider. This is intended for future use where other
   data might need to be cached, and it prevents a lot of re-wiring.

5. A different caching implementation may be set in the registry
   when initializing Vulcan (just like plugins), and this implementation,
   when non-nil, will be used in place of the Engine-based caches.

   This is very useful when testing vulcan itself and might require
   rapid restarts with the Memory Engine. Requesting the same
   certificate from Let's Encrypt repeatedly will easily
   get you banned.

   In this case, it might be worthwhile to use the DirCache implementation
   which caches on disk, instead of Etcd.

   This also enables potential caching to happen on remote/reliable
   KV stores such as Redis, DynamoDB, etc.

Cosmetic changes + UT for Caching

Lock memory-based autocert cache with a Mutex

SNI doesn't work

Switch to dep vendoring tool

Build with go 1.9 (fixes #341)

Fix SNI to work with multiple hosts correctly

When testing AutoCert, I realized that this test was not comprehensive enough.

If we add an additional host call, say "foobar", we get a 404 back because the
route doesn't exist for it. However, the default cert is returned regardless.

This means that if we checked for the Certificate Serial Number on the returned
cert, it is constant across any host name. In short, exposing multiple SSL
Certs by Host, will not result in different certs being returned
based on the request made.

I have set up the initial conditions for the test which fails right now.

I'll figure out how to make it pass.

Fixed SNI by setting client-side ServerName in each TLS request.

Fixed SNI by setting client-side ServerName in each TLS request.

Updated SNI test to be better formed

Cleaned up UTs

Gofmt + goimports

* Removed unnecessary testing in SNI, and lowercased the getPeerCertSerialNo

* Add space after every // globally

* Simplified the newTLSCfg function

1. Broken the newTlsCfg function into sub-functions to make it
   easily readable and understandbale. Each function is a few lines
   long, with few conditionals and easy to understand.

2. This allowed me to perform OCSP stapling even on AutoCert
   generated certificates by calling into a function.

* Removed references to server in ocspstaple function

* Allow for OCSP Stapling when AutoCert'ing

* Added RSA and ECDSA Client ID key tests

* Verify client-ID key passed successfully from host-settings

* go fmt

* Implemented ocspStapleToCert in a fail-fast/return-fast style